requirements.txt
flask
opencv-python
pillow
numpy
onnxruntime
ultralytics
